requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
attrs==25.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dataclasses-json==0.6.7
DateTime==5.5
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
gotrue==2.12.0
greenlet==3.1.1
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.51
langchain-openai==0.3.12
langchain-pinecone==0.2.5
langchain-tests==0.3.17
langchain-text-splitters==0.3.8
langsmith==0.3.28
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.3.2
mypy-extensions==1.0.0
narwhals==1.34.0
numpy==2.2.4
openai==1.70.0
orjson==3.10.16
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pinecone==6.0.2
pinecone-plugin-interface==0.0.7
pluggy==1.5.0
postgrest==1.0.1
propcache==0.3.1
protobuf==5.29.4
pyarrow==19.0.1
pydantic==2.11.2
pydantic-settings==2.8.1
pydantic_core==2.33.1
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pypdf==5.4.0
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-mock==3.14.0
pytest-socket==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
realtime==2.4.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.1
rpds-py==0.24.0
setuptools==78.1.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
starlette==0.46.1
storage3==0.11.3
streamlit==1.44.1
StrEnum==0.4.15
supafunc==0.9.4
syrupy==4.9.1
tenacity==9.1.2
tiktoken==0.9.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.0.4
websockets==14.2
yarl==1.19.0
zope.interface==7.2
zstandard==0.23.0
